Move utility into Utilities namespace.
