+++ src/Aquifer.API/Endpoints/Comments/Create/Endpoint.cs
@@ -2,7 +2,7 @@
 using Aquifer.API.Services;
 using Aquifer.Data;
 using Aquifer.Data.Entities;
-using Aquifer.Data.Services;
+using Aquifer.Data.Utilities;
 using FastEndpoints;
 using Microsoft.EntityFrameworkCore;
