+++ src/Aquifer.API/Endpoints/Comments/Threads/Create/Endpoint.cs
@@ -4,7 +4,7 @@ using Aquifer.API.Helpers;
 using Aquifer.API.Services;
 using Aquifer.Data;
 using Aquifer.Data.Entities;
-using Aquifer.Data.Services;
+using Aquifer.Data.Utilities;
 using FastEndpoints;
 using Microsoft.EntityFrameworkCore;
